use flow with Room to replace manual management
Removed the need to predict where will reloading of the known item id's
be needed to avoid errors. The collection of IDs is assumed to
always contain up to date values.

One issue with that assumption is that initially, it's not the case, and
in theory this could lead to race conditions, like it would in the init
block of the PersistenceViewModel, where loadNext() called right after
setting up the flow, can run before the values are obtained from the
database.