fix MutableCyclicIterator ignoring new last value
On first pass, the new last value was being  be ignored after replaceData()
call, because the implementation moved the currentItem index after
next() call, resetting it effectively when the last item was reached.

That corner-case was fixed by first incrementing the index and then
using it to pick the value to return in next().

Also added a test to guard against future regressions.