fix concurrency issues in MutableSingleFlowEvent
A stress test was added to check how the single event behaves when
multiple observers are present and a lot of events are sent.
The former implementation had issues with delay between collect call and
the start of the flow inside the class, thus, under heavy load, it was
sometimes losing events.

New implementation is based on a SharedFlow, it's potential limitation
is the coroutineScope it creates inside itself.