add cancellation post exception sample
Even exceptions handled with coroutineExceptionhandlers by us, not by
the default one which just crashes the app in case of an exception,
still propagate cancellation. Even though we might run some custom
action when we catch an exception using the handler - that's more
useful for some cleanup, won't stop the cancellation on its own.