add parallel flows experiment
added experimental code in flows screen's logic, trying emissions to a
flow which handles them in parallel and outputs the number of the
request last processed.

Also enhanced other parallel flows samples with limitation of their
concurrency to the available processors count. This future-proofs the
code against running on >16 core machines, as 16 is the default
concurrency for flatMapMerge.