fix network exception sample
The exception was propagated and cancelled the entire scope for good, so
no later sample runs could be sucessful. Now a supervisor job is added
one level above the launch in which the exception is thrown.

Note: usually on android, the scopes provided by ktx extensions would
have supervisor job baked-in already, so this would not be needed, here
however, a custom scope is used, which intentionally does have a regular
Job instead, so wider variety of samples can be presented.